Move stylesheets to styles folder, update refs
