+++ src/styles/hero.module.css
@@ -5,7 +5,6 @@
   color: white;
   min-height: 90vh;
 }
-
 .name {
   line-height: 2em;
   box-sizing: border-box;
@@ -20,6 +19,7 @@
 }
 .content {
   display: flex;
+  width: 100%;
   flex-direction: column;
   justify-content: center;
   align-items: center;
